feat: 支持指令删除插件 (#286)
